Move TaskList class from util to tasks package
